Skip file cleanup if the tmp directory has not been created
- This issue presents when running the test suite for the first time
  (e.g. in CI) since there is no `tests/_files/tmp` directory created.
- In the one test case which requires the cleanup of these files because
  it creates them it does not fail because the directory is created.
- The other test cases don't need this tear down, but they fail when it
  tries to run as this directory has not yet been created.
- In future we'll probably remove both the tear down and the test case
  that needs it as this is probably not an appropriately direct way to
  test this.

+++ tests/tests/Data/ProcessedCodeCoverageDataMapperTest.php
@@ -32,7 +32,11 @@ final class ProcessedCodeCoverageDataMapperTest extends TestCase
     {
         parent::tearDown();
 
-        foreach (new FilesystemIterator(self::$TEST_TMP_PATH) as $fileInfo) {
+        if (!file_exists(self::$TEST_TMP_PATH)) {
+            return;
+        }
+
+        foreach (new FilesystemIterator(self::$TEST_TMP_PATH) as $path => $fileInfo) {
             /* @var \SplFileInfo $fileInfo */
             unlink($fileInfo->getPathname());
         }
